+++ app/helpers.php
@@ -3,7 +3,7 @@ declare (strict_types = 1);
 
 use My\Theme;
 
-function Theme(): Theme
+function MyTheme(): Theme
 {
     return Theme::getInstance();
 }