+++ app/libraries/Theme/Setups/Scripts/CustomizePreview.php
@@ -21,7 +21,7 @@ final class CustomizePreview extends AbstractScript
     {
         parent::__construct($theme);
 
-        $this->id = 'jentil-theme-customize-preview';
+        $this->id = "{$this->app->theme->stylesheet}-customize-preview";
     }
 
     public function run()
@@ -66,14 +66,15 @@ final class CustomizePreview extends AbstractScript
      */
     public function enqueue()
     {
+        $file = '/dist/scripts/customize-preview.min.js';
+
         \wp_enqueue_script(
             $this->id,
-            $this->app->utilities->fileSystem->themeDir(
-                'url',
-                '/dist/scripts/customize-preview.min.js'
-            ),
+            $this->app->utilities->fileSystem->themeDir('url', $file),
             ['customize-preview'],
-            '',
+            \filemtime(
+                $this->app->utilities->fileSystem->themeDir('path', $file)
+            ),
             true
         );
     }
@@ -83,7 +84,7 @@ final class CustomizePreview extends AbstractScript
      */
     public function addInlineScript()
     {
-        $script = 'var jentilThemeAwesomePostsHeadingModId = "'.
+        $script = 'var myThemeAwesomePostsHeadingModId = "'.
             $this->app->setups['Customizer']
                 ->sections['AwesomePosts']
                 ->settings['Heading']->id.