+++ functions.php
@@ -4,47 +4,47 @@
  * NOTE: Keep code in this file compatible with PHP 5.2
  */
 
-define('THEME_SLUG', 'my-theme'); // Update as appropriate
-define('MINIMUM_PHP', '7.0');
-define('MINIMUM_WP', '4.7');
+define('MY_THEME_MIN_PHP', '7.0');
+define('MY_THEME_MIN_WP', '4.7');
 
-if (version_compare(PHP_VERSION, MINIMUM_PHP, '<') ||
-    version_compare(get_bloginfo('version'), MINIMUM_WP, '<')
+if (version_compare(PHP_VERSION, MY_THEME_MIN_PHP, '<') ||
+    version_compare(get_bloginfo('version'), MY_THEME_MIN_WP, '<')
 ) {
-    add_action('admin_notices', 'printJentilThemeReqNotice');
+    add_action('admin_notices', 'printMyThemeReqNotice');
 
-    deactivateJentilTheme();
+    deactivateMyTheme();
 } else {
     require __DIR__.'/vendor/autoload.php';
 
-    add_action('after_setup_theme', 'runJentilTheme', 2);
+    add_action('after_setup_theme', 'runMyTheme', 2);
 }
 
-function runJentilTheme()
+function runMyTheme()
 {
-    Theme()->run();
+    MyTheme()->run();
 }
 
-function printJentilThemeReqNotice()
+function printMyThemeReqNotice()
 {
     echo '<div class="notice notice-error">
         <p>'.
         sprintf(
-            __(
-                '%1$s theme has been deactivated as it requires PHP >= %2$s and WordPress >= %3$s'
+            esc_html__(
+                '%1$s theme has been deactivated as it requires PHP >= %2$s and WordPress >= %3$s',
+                'my-theme'
             ),
-            '<code>'.THEME_SLUG.'</code>',
-            '<strong>'.MINIMUM_PHP.'</strong>',
-            '<strong>'.MINIMUM_WP.'</strong>'
+            '<code>my-theme</code>',
+            '<strong>'.MY_THEME_MIN_PHP.'</strong>',
+            '<strong>'.MY_THEME_MIN_WP.'</strong>'
         ).
         '</p>
     </div>';
 }
 
-function deactivateJentilTheme()
+function deactivateMyTheme()
 {
     $themes = wp_get_themes(['allowed' => true]);
-    unset($themes[THEME_SLUG]);
+    unset($themes['my-theme']);
 
     $theme = reset($themes);
     $name = null === key($themes) ? '' : $theme->get_stylesheet();