+++ tests/unit/libraries/Theme/Setups/Translations/CoreTest.php
@@ -1,7 +1,7 @@
 <?php
 declare (strict_types = 1);
 
-namespace My\Theme\Setups;
+namespace My\Theme\Setups\Translations;
 
 use My\Theme\AbstractTestCase;
 use My\Theme\Utilities;
@@ -10,21 +10,28 @@ use GrottoPress\Jentil\AbstractChildTheme;
 use Codeception\Util\Stub;
 use tad\FunctionMocker\FunctionMocker;
 
-class LanguageTest extends AbstractTestCase
+class CoreTest extends AbstractTestCase
 {
     public function testRun()
     {
         $add_action = FunctionMocker::replace('add_action');
 
-        $language = new Language(Stub::makeEmpty(AbstractChildTheme::class));
+        $translation = new Core(Stub::makeEmpty(AbstractChildTheme::class, [
+            'theme' => new class {
+                function get()
+                {
+                    return 'my-theme';
+                }
+            }
+        ]));
 
-        $language->run();
+        $translation->run();
 
         $add_action->wasCalledOnce();
 
         $add_action->wasCalledWithOnce([
             'after_setup_theme',
-            [$language, 'loadTextDomain']
+            [$translation, 'loadTextDomain']
         ]);
     }
 
@@ -34,22 +41,29 @@ class LanguageTest extends AbstractTestCase
 
         $theme = Stub::makeEmpty(AbstractChildTheme::class, [
             'utilities' => Stub::makeEmpty(Utilities::class),
+            'theme' => new class {
+                function get()
+                {
+                    return 'my-theme';
+                }
+            }
         ]);
 
+        $theme->utilities = Stub::makeEmpty(Utilities::class);
         $theme->utilities->fileSystem = Stub::makeEmpty(FileSystem::class, [
             'themeDir' => function (string $type, string $append) {
                 return "/var/www/themes/my-theme{$append}";
             }
         ]);
 
-        $language = new Language($theme);
+        $translation = new Core($theme);
 
-        $language->loadTextDomain();
+        $translation->loadTextDomain();
 
         $load->wasCalledOnce();
         $load->wasCalledWithOnce([
-            'my-theme',
-            '/var/www/themes/my-theme/languages'
+            $translation->textDomain,
+            '/var/www/themes/my-theme/lang'
         ]);
     }
 }